Improved handling of situations in which the three axes cannot move over the same duration.
At a high level, I find that panning and tilting is less visually distracting when at a wider setting.
So where possible, camera motion should occur while the camera is zoomed out as much as possible.  Thus,
when you have a recall that includes some combination of panning, tilting, and zooming, if the axes
cannot stop at the same time, this code now prefers to pan and tilt either before zooming in or at the
beginning of zooming in, but after or at the end of zooming out.

The rules are as follows:

  * If the duration of panning or tilting is less than the total move AND the duration of zooming is LONGER than
    the pan or tilt duration:

      - If the camera is zooming in, perform the pan or tilt immediately, before the zoom goes in too far.
      - If the camera is zooming out, defer the pan or tilt so that the operations end at the same time.

  * If the zoom duration is shorter than the total move duration:

      - If the camera is zooming out, zoom out immediately, aligning it with the beginning of the longer
        pan or tilt move.
      - If the camera is zooming in, defer the zoom so that the operations end at the same time.

This should result in performing pan and tilt operations when the zoom is as wide as possible given the
provided durations.  It may, under certain circumstances, cause pan and tilt operations to start at
different times.  This is intentional.