Improved the way that recall durations are chosen.
The old code used fixed durations, where fast moves were a small
number of seconds and slow moves were up to 30 seconds, then
adjusted them as needed to fit within the maximum or minimum
durations that are achievable by hardware.  This approach was
okay, but wasn't great for short moves.

The new code uses the maximum and minimum physically possible
durations as guides, and chooses a duration such that the maximum
speed is close to the minimum physically possible duration and
vice versa, with a cap of 30 seconds.  This seems like a more
reasonable approach.