Added preliminary 3D printing files for a Bescor MP-101 encoder mod.
This mod adds absolute positioning capabilities to the MP-101.  Note that
this hardware is entirely untested, but it has been successfully assembled.

This mod removes the battery compartment, because powering the required
electronics from AA batteries would not be feasible.  It also makes the
hardware slightly larger.

This is not supported at all in software yet.

Details in README.md.